--- src/cli/commands/index.ts
@@ -1,1 +1,0 @@
-export * from './entropy-program'
